+++ contracts/counter/src/contract.rs
@@ -1,5 +1,5 @@
 use cosmwasm_std::{
-    entry_point, to_binary, Binary, Deps, DepsMut, Env, MessageInfo, Response, StdResult,
+    entry_point, to_json_binary, Binary, Deps, DepsMut, Env, MessageInfo, Response, StdResult,
 };
 use cw2::set_contract_version;
 
@@ -48,7 +48,7 @@ pub fn execute(
 #[cfg_attr(feature = "export", entry_point)]
 pub fn query(deps: Deps, _env: Env, msg: QueryMsg) -> StdResult<Binary> {
     match msg {
-        QueryMsg::GetCount {} => to_binary(&query::count(deps)?),
+        QueryMsg::GetCount {} => to_json_binary(&query::count(deps)?),
     }
 }
 
@@ -62,7 +62,7 @@ pub fn migrate(_deps: DepsMut, _env: Env, _msg: MigrateMsg) -> Result<Response, 
 mod tests {
     use super::*;
     use cosmwasm_std::{
-        coins, from_binary,
+        coins, from_json,
         testing::{mock_dependencies, mock_env, mock_info},
     };
 
@@ -79,7 +79,7 @@ mod tests {
 
         // it worked, let's query the state
         let res = query(deps.as_ref(), mock_env(), QueryMsg::GetCount {}).unwrap();
-        let value: GetCountResponse = from_binary(&res).unwrap();
+        let value: GetCountResponse = from_json(res).unwrap();
         assert_eq!(17, value.count);
     }
 
@@ -98,7 +98,7 @@ mod tests {
 
         // should increase counter by 1
         let res = query(deps.as_ref(), mock_env(), QueryMsg::GetCount {}).unwrap();
-        let value: GetCountResponse = from_binary(&res).unwrap();
+        let value: GetCountResponse = from_json(res).unwrap();
         assert_eq!(18, value.count);
     }
 
@@ -126,7 +126,7 @@ mod tests {
 
         // should now be 5
         let res = query(deps.as_ref(), mock_env(), QueryMsg::GetCount {}).unwrap();
-        let value: GetCountResponse = from_binary(&res).unwrap();
+        let value: GetCountResponse = from_json(res).unwrap();
         assert_eq!(5, value.count);
     }
 }