+++ contracts/mock_contract_u64/src/lib.rs
@@ -1,7 +1,8 @@
 use mock_contract::{ExecuteMsg, InstantiateMsg, MigrateMsg, QueryMsg};
 
 use cosmwasm_std::{
-    entry_point, to_binary, Binary, Deps, DepsMut, Env, MessageInfo, Response, StdError, StdResult,
+    entry_point, to_json_binary, Binary, Deps, DepsMut, Env, MessageInfo, Response, StdError,
+    StdResult,
 };
 
 #[entry_point]
@@ -54,10 +55,10 @@ pub fn execute(
 #[entry_point]
 pub fn query(_deps: Deps, _env: Env, msg: QueryMsg<u64>) -> StdResult<Binary> {
     match msg {
-        QueryMsg::FirstQuery {} => to_binary("first query passed"),
+        QueryMsg::FirstQuery {} => to_json_binary("first query passed"),
         QueryMsg::SecondQuery { .. } => Err(StdError::generic_err("Query not available")),
-        QueryMsg::ThirdQuery { .. } => to_binary("third query passed"),
-        QueryMsg::FourthQuery(_, _) => to_binary("fourth query passed"),
+        QueryMsg::ThirdQuery { .. } => to_json_binary("third query passed"),
+        QueryMsg::FourthQuery(_, _) => to_json_binary("fourth query passed"),
     }
 }