+++ cw-orch-daemon/tests/common/mod.rs
@@ -7,11 +7,13 @@ mod node {
 
     use ctor::{ctor, dtor};
 
-    use cw_orch_core::env::CwOrchEnvVars;
+    use cw_orch_core::CwOrchEnvVars;
     use duct::cmd;
 
     // Config
     const JUNO_IMAGE: &str = "ghcr.io/cosmoscontracts/juno:v12.0.0";
+    #[allow(unused)]
+    pub const STAKE_TOKEN: &str = "ujunox";
 
     // Defaults for env vars
     const CONTAINER_NAME: &str = "juno_node_1";
@@ -47,6 +49,8 @@ mod node {
     }
 
     pub mod container {
+        use crate::common::STAKE_TOKEN;
+
         use super::cmd;
 
         pub fn find(name: &String) -> bool {
@@ -85,7 +89,7 @@ mod node {
                 "-p",
                 "9090:9090",
                 "-e",
-                "STAKE_TOKEN=ujunox",
+                format!("STAKE_TOKEN={}", STAKE_TOKEN),
                 "-e",
                 "UNSAFE_CORS=true",
                 image,
@@ -149,14 +153,7 @@ mod node {
         }
         let image = env::var("JUNO_IMAGE").unwrap();
 
-        let temp_dir = env::temp_dir();
-        let state_file = temp_dir.join("cw_orch_test.json");
-
-        if CwOrchEnvVars::StateFile.get().is_err() {
-            env::set_var("STATE_FILE", state_file);
-        }
-
-        if CwOrchEnvVars::LocalMnemonic.get().is_err() {
+        if CwOrchEnvVars::load().unwrap().local_mnemonic.is_none() {
             env::set_var("LOCAL_MNEMONIC", LOCAL_MNEMONIC);
         }
 
@@ -164,11 +161,11 @@ mod node {
         log::info!("Using CONTAINER_NAME: {}", container);
         log::info!(
             "Using STATE_FILE: {}",
-            CwOrchEnvVars::StateFile.get().unwrap()
+            CwOrchEnvVars::load().unwrap().state_file.display()
         );
         log::info!(
-            "Using LOCAL_MNEMONIC: {}",
-            CwOrchEnvVars::LocalMnemonic.get().unwrap()
+            "Using LOCAL_MNEMONIC: {:?}",
+            CwOrchEnvVars::load().unwrap().local_mnemonic
         );
 
         container::start(&container, &image);