+++ packages/cw-orch-core/src/contract/paths.rs
@@ -58,8 +58,8 @@ mod wasm_path {
 mod artifacts_dir {
     use super::WasmPath;
     use crate::{
-        build::BuildPostfix, env::CwOrchEnvVars, environment::ChainState, error::CwEnvError,
-        log::LOCAL_LOGS,
+        build::BuildPostfix, env::ARTIFACTS_DIR_ENV_NAME, environment::ChainState,
+        error::CwEnvError, log::LOCAL_LOGS, CwOrchEnvVars,
     };
 
     use std::{env, fs, path::PathBuf};
@@ -115,9 +115,10 @@ mod artifacts_dir {
     impl ArtifactsDir {
         /// Get the artifacts directory from the environment variable `ARTIFACTS_DIR`.
         pub fn env() -> Self {
-            let dir = CwOrchEnvVars::ArtifactsDir
-                .get()
-                .unwrap_or_else(|_| panic!("{} env variable not set", CwOrchEnvVars::ArtifactsDir));
+            let dir = CwOrchEnvVars::load()
+                .unwrap()
+                .artifacts_dir
+                .unwrap_or_else(|| panic!("{} env variable not set", ARTIFACTS_DIR_ENV_NAME));
             Self::new(dir)
         }