+++ publish.sh
@@ -45,6 +45,8 @@ done
 
 echo "Everything is published!"
 
-VERSION=$(< Cargo.toml grep -m 1 version | sed 's/-/_/g' | grep -o '".*"' | sed 's/"//g');
-git tag v"$VERSION"
-git push origin v"$VERSION"
+# 
+# VERSION=$(< Cargo.toml grep -m 1 version | sed 's/-/_/g' | grep -o '".*"' | sed 's/"//g');
+# git push
+# git tag v"$VERSION"
+# git push origin v"$VERSION"